requirements.txt
flet==0.12.1
hidapi==0.14.0
python-osc==1.8.1
requests==2.31.0
